[Modify] step1 controller 실행 메서드 추상화 레벨 수정
